% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{add_layout}
\alias{add_layout}
\title{Add coordinates based on an igraph layout}
\usage{
add_layout(x, layout, seed = 1234, xmin = 0, xmax = 1200, ymin = 0,
  ymax = 800, ...)
}
\arguments{
\item{x}{object of class visjs}

\item{layout}{layout function or string}

\item{seed}{random seed to initialize prior to computing the layout}

\item{xmin}{minimum value of x to scale the layout to}

\item{xmax}{maximum value of x to scale the layout to}

\item{ymin}{minimum value of y to scale the layout to}

\item{ymax}{maximum value of y to scale the layout to}

\item{...}{additional arguments to pass to the layout function}
}
\description{
Add coordinates based on an igraph layout
}

