% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/c3.R
\name{c3}
\alias{c3}
\title{Create a plot using C3.js}
\usage{
c3(..., width = NULL, height = NULL)
}
\arguments{
\item{...}{payload for C3 plot}

\item{width}{width of the plot}

\item{height}{height of the plot}
}
\description{
Create a plot using C3.js
}

