% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keendashboard.R
\name{Column}
\alias{Column}
\title{Create a column}
\usage{
Column(width, ..., offset = NULL, id = paste0("column-", round(runif(1) *
  10000)))
}
\arguments{
\item{width}{the grid width of the column}

\item{...}{elements to include within the column}

\item{offset}{the number of columns to offset this column}

\item{id}{id for the column}
}
\description{
Create a column
}

