% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dashboards.R
\name{dtDashboard}
\alias{dtDashboard}
\title{Network visualization dashboard}
\usage{
dtDashboard(p1, p4, p5)
}
\arguments{
\item{p1}{object of class visjs}

\item{p4}{object for panel 2}

\item{p5}{object for panel 3}
}
\description{
Network visualization dashboard
}

