% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extras.R
\name{dyAlteryxTheme}
\alias{dyAlteryxTheme}
\title{Customize dygraph theme for Alteryx Dashboards}
\usage{
dyAlteryxTheme(x)
}
\arguments{
\item{x}{dygraph object}
}
\description{
Customize dygraph theme for Alteryx Dashboards
}

