% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{group_by}
\alias{group_by}
\title{Cluster nodes based on a variable or a function}
\usage{
group_by(x, f, ...)
}
\arguments{
\item{x}{object of class visjs}

\item{f}{function or variable to use for clustering nodes}

\item{...}{additional arguments to pass to function f}
}
\description{
Cluster nodes based on a variable or a function
}

