% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/horizontalBarChart.R
\name{horizontalBarChart}
\alias{horizontalBarChart}
\title{Plot an interactive horizontal bar chart}
\usage{
horizontalBarChart(data, ..., width = 400, height = 400)
}
\arguments{
\item{data}{data to plot}

\item{...}{additional arguments to pass as defaults}

\item{width}{width of the plot div}

\item{height}{height of the plot div}
}
\description{
Plot an interactive horizontal bar chart
}

