% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hovercard.R
\name{hovercard}
\alias{hovercard}
\title{Create a hovercard}
\usage{
hovercard(text, details, width = NULL, height = NULL)
}
\arguments{
\item{text}{text to hover on}

\item{details}{details to display on hover}

\item{width}{width of the element}

\item{height}{height of the element}
}
\description{
Create a hovercard
}

