% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iacf.R
\name{iacf}
\alias{iacf}
\title{Create an interactive ACF plot}
\usage{
iacf(data, hlines, ydomain, xlab = "Lag", ylab = "ACF", width = NULL,
  height = NULL)
}
\arguments{
\item{data}{data to plot}

\item{hlines}{horizontal markers to plot}

\item{ydomain}{range for the y-axis}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{width}{width of the plot}

\item{height}{height of the plot}
}
\description{
Create an interactive ACF plot
}

