% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iacf.R
\name{iacfOutput}
\alias{iacfOutput}
\title{Widget output function for use in Shiny}
\usage{
iacfOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{id of output element}

\item{width}{width of widget div}

\item{height}{height of widget div}
}
\description{
Widget output function for use in Shiny
}

