% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iConfusionMatrix.R
\name{iConfusionMatrix}
\alias{iConfusionMatrix}
\title{Plot an Interactive Confusion Matrix}
\usage{
iConfusionMatrix(data, ..., width = 400, height = 400)
}
\arguments{
\item{data}{confusion matrix data}

\item{...}{additional arguments to pass}

\item{width}{width of the plot div}

\item{height}{height of the plot div}
}
\description{
Plot an Interactive Confusion Matrix
}

