% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ihist.R
\name{ihist}
\alias{ihist}
\title{Create an interactive histogram}
\usage{
ihist(data, ..., width = NULL, height = NULL)
}
\arguments{
\item{data}{list of vectors}

\item{...}{extra arguments}

\item{width}{width of the histogram}

\item{height}{height of the histogram}
}
\description{
Create an interactive histogram
}

