% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infobox.R
\name{infobox}
\alias{infobox}
\title{Create an infobox.}
\usage{
infobox(..., div, width = NULL, height = NULL)
}
\arguments{
\item{...}{items in the infobox}

\item{div}{div element for the infobox}

\item{width}{width of the box}

\item{height}{height of the box}
}
\description{
Create an infobox.
}

