% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iplot.R
\name{iplot.acf}
\alias{iplot.acf}
\title{Create interactive plot of an object of class acf}
\usage{
\method{iplot}{acf}(x, width = "100\%", height = 300, ...)
}
\arguments{
\item{x}{an object of class forecast}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{...}{other arguments to pass to dygraphs::dygraph}
}
\description{
Create interactive plot of an object of class acf
}

