% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iplot.R
\name{iplot.forecast}
\alias{iplot.forecast}
\title{Create interactive plot of an object of class forecast}
\usage{
\method{iplot}{forecast}(x, width = NULL, height = NULL, ...)
}
\arguments{
\item{x}{an object of class forecast}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{...}{other arguments to pass to dygraphs::dygraph}
}
\description{
Create interactive plot of an object of class forecast
}
\examples{
library(forecast)
 fit <- ets(USAccDeaths)
 iplot(forecast(fit, h = 48))
}

