% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iplot.R
\name{iplot}
\alias{iplot}
\title{Generic function for creating interactive plots for time series objects}
\usage{
iplot(x, width, height, ...)
}
\arguments{
\item{x}{a time series object}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{...}{other arguments for the plot}
}
\description{
Generic function for creating interactive plots for time series objects
}

