% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iplot.R
\name{iplot.stl}
\alias{iplot.stl}
\title{Create interactive plot of an object of class stl}
\usage{
\method{iplot}{stl}(x, width = 500, height = 120, labelsDiv = NULL, ...)
}
\arguments{
\item{x}{an object of class ets}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{labelsDiv}{div element to display legend of dygraph}

\item{...}{other arguments to pass to dygraphs::dygraph}
}
\description{
Create interactive plot of an object of class stl
}

