% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iplot.R
\name{iplot.ts}
\alias{iplot.ts}
\title{Create interactive plot of an object of class ts}
\usage{
\method{iplot}{ts}(x, width = "100\%", height = 300, label = "V1", ...)
}
\arguments{
\item{x}{an object of class ts}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{label}{label to use for the y variable in legend}

\item{...}{other arguments to pass to dygraphs::dygraph}
}
\description{
Create interactive plot of an object of class ts
}

