% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keendashboard.R
\name{keen_dash}
\alias{keen_dash}
\title{Create a dashboard}
\usage{
keen_dash(...)
}
\arguments{
\item{...}{components of dashboard}
}
\description{
This function creates a dashboard using a framework provided by
\url{https://github.com/keen/dashboards}
}

