% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keendashboard.R
\name{Navbar}
\alias{Navbar}
\title{Create a navbar at the top of the page}
\usage{
Navbar(brand, ...)
}
\arguments{
\item{brand}{title to display on far left of navbar}

\item{...}{other elements in navbar}
}
\description{
Create a navbar at the top of the page
}

