% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dashboards.R
\name{networkVizDashboard}
\alias{networkVizDashboard}
\title{Network visualization dashboard}
\usage{
networkVizDashboard(p1)
}
\arguments{
\item{p1}{object of class visjs}
}
\description{
Network visualization dashboard
}

