% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iplot.R
\name{plotData.acf}
\alias{plotData.acf}
\title{Extract plot data from an object of class acf}
\usage{
\method{plotData}{acf}(x)
}
\arguments{
\item{x}{an object of class acf}
}
\description{
Extract plot data from an object of class acf
}
\examples{
library(forecast)
x = Acf(wineind, plot = FALSE)
plotData(x)
}

