% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iplot.R
\name{plotData.ets}
\alias{plotData.ets}
\title{Extract plot data from an object of class ets}
\usage{
\method{plotData}{ets}(x)
}
\arguments{
\item{x}{an object of class ets}
}
\description{
Extract plot data from an object of class ets
}

