% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iplot.R
\name{plotData.forecast}
\alias{plotData.forecast}
\title{Extract plot data from an object of class forecast}
\usage{
\method{plotData}{forecast}(x)
}
\arguments{
\item{x}{an object of class forecast}
}
\description{
Extract plot data from an object of class forecast
}

