% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iplot.R
\name{plotData}
\alias{plotData}
\title{Generic function for extracting plot data from time series objects.}
\usage{
plotData(x)
}
\arguments{
\item{x}{a time series object}
}
\description{
Generic function for extracting plot data from time series objects.
}

