% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iplot.R
\name{plotData.stl}
\alias{plotData.stl}
\title{Extract plot data from an object of class stl}
\usage{
\method{plotData}{stl}(x)
}
\arguments{
\item{x}{an object of class stl}
}
\description{
Extract plot data from an object of class stl
}

