% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/renderInComposer.R
\name{renderInComposer}
\alias{renderInComposer}
\title{Render a htmlwidget in Alteryx Composer}
\usage{
renderInComposer(x, libdir = NULL, nOutput = 3, debug = FALSE)
}
\arguments{
\item{x}{an object of class htmlwidget or shiny.tag}

\item{libdir}{directory to copy js/css dependencies to}

\item{nOutput}{node to pipe output back into Alteryx}

\item{debug}{only for use in development}
}
\description{
Render a htmlwidget in Alteryx Composer
}

