% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infobox.R
\name{renderInfobox}
\alias{renderInfobox}
\title{Widget render function for use in Shiny}
\usage{
renderInfobox(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{A quoted of unquoted expression, or a function}

\item{env}{The desired environment for the function. Defaults to the calling
environment two steps back}

\item{quoted}{Is the expression quoted?}
}
\description{
Widget render function for use in Shiny
}

