% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tablechart.R
\name{renderTablechart}
\alias{renderTablechart}
\title{Widget render function for use in Shiny}
\usage{
renderTablechart(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

