% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trees.R
\name{renderTree}
\alias{renderTree}
\title{Render Decision Tree}
\usage{
renderTree(fit, colpal = NULL, tooltipParams = NULL)
}
\arguments{
\item{fit}{model object of class rpart}

\item{colpal}{color palette}

\item{tooltipParams}{extra parameter to pass to tooltip}
}
\description{
Render Decision Tree
}

