% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{size_by}
\alias{size_by}
\title{Size vertices based on a variable or a function}
\usage{
size_by(x, f, scaling = list(min = 10, max = 30), shape = "dot", ...)
}
\arguments{
\item{x}{object of class visjs}

\item{f}{function or variable to use for scaling nodes}

\item{scaling}{named list of min and max values to use for scaling}

\item{shape}{string indicating the shape of the node}

\item{...}{additional arguments to pass to the function f}
}
\description{
Size vertices based on a variable or a function
}

