% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tablechart.R
\name{tablechart}
\alias{tablechart}
\title{Create a table chart}
\usage{
tablechart(measures, heading = "Measure", models = "", footnote = NULL,
  width = NULL, height = NULL)
}
\arguments{
\item{measures}{measures to plot}

\item{heading}{table heading}

\item{models}{model names}

\item{footnote}{footnote to include}

\item{width}{width of the plot div}

\item{height}{of the plot div}
}
\description{
This widget creates a sortable table chart.
}

