% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{update_options}
\alias{update_options}
\title{Update options}
\usage{
update_options(x, ...)
}
\arguments{
\item{x}{object of class visjs}

\item{...}{named options}
}
\description{
Update options
}

