% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trees.R
\name{varImpPlot}
\alias{varImpPlot}
\title{Interactive Variable Importance Plot}
\usage{
varImpPlot(fit, limit = 10, ...)
}
\arguments{
\item{fit}{model object of class rpart}

\item{limit}{number of variables to plot}

\item{...}{additional arguments}
}
\description{
Interactive Variable Importance Plot
}

