% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vis.R
\name{vis.data.frame}
\alias{vis.data.frame}
\title{Create a network plot from a data frame}
\usage{
\method{vis}{data.frame}(x, options = c(), directed = FALSE,
  vertices = NULL, type = "edgelist", width = NULL, height = NULL,
  elementId = NULL, ...)
}
\arguments{
\item{x}{data frame of edges}

\item{options}{named list of options}

\item{directed}{boolean indicating if the graph is directed}

\item{vertices}{optional data frame of vertices}

\item{type}{indicating nature of edge data frame - edgelist of adj. matrix.}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{elementId}{id of the widget}

\item{...}{additional arguments}
}
\description{
Create a network plot from a data frame
}

