% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vis.R
\name{vis.igraph}
\alias{vis.igraph}
\title{Create network plot from an igraph object}
\usage{
\method{vis}{igraph}(x, options = c(), ..., width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{x}{object of class igraph to create network plot from}

\item{options}{named list of options}

\item{...}{additional arguments}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{elementId}{id of the widget}
}
\description{
Create network plot from an igraph object
}

